+++ src/sensors/BoschBMP3xx.cpp
@@ -105,7 +105,6 @@ bool BoschBMP3xx::setup(void) {
                  "recommended"));
     }
 
-
     // convert the standby time enum value into the time between readouts from
     // the BMP's
     // ADC NOTE:  The ADC will return repeated values if the ADC's ODR (output
@@ -163,33 +162,39 @@ bool BoschBMP3xx::setup(void) {
                  "trim parameters"));
         success = bmp_internal.begin(_i2cAddressHex);
 
-        // Set up oversampling and filter initialization
-        // Using the filter selection recommended for "Weather monitoring
-        // (lowest power)" in table 10 of the sensor datasheet
+        if (success) {
+            // The sea level pressure is used for calculating altitude. This is
+            // stored as a parameter of the bmp_internal object and only needs
+            // to be sent once at setup, not repeated at wake, even if we're not
+            // continuously powered.
+            MS_DBG(F("Setting sea level atmospheric pressure to"),
+                   MS_SEA_LEVEL_PRESSURE_HPA);
+            bmp_internal.setSeaLevelPressure(MS_SEA_LEVEL_PRESSURE_HPA);
 
-        // Oversampling setting
-        MS_DBG(F("Sending BMP3xx oversampling settings"));
-        bmp_internal.setTempOversampling(_tempOversampleEnum);
-        bmp_internal.setPresOversampling(_pressureOversampleEnum);
+            // Oversampling settings - these settings are sent to the BMP3xx and
+            // need to be repeated at wake if we're not continuously powered
+            MS_DBG(F("Sending BMP3xx oversampling settings"));
+            bmp_internal.setTempOversampling(_tempOversampleEnum);
+            bmp_internal.setPresOversampling(_pressureOversampleEnum);
 
-        // Coefficient of the filter (in samples)
-        MS_DBG(F("Sending BMP3xx IIR Filter settings"));
-        bmp_internal.setIIRFilter(_filterCoeffEnum);
-
-        MS_DBG(F("Setting sea level atmospheric pressure to"),
-               MS_SEA_LEVEL_PRESSURE_HPA);
-        bmp_internal.setSeaLevelPressure(MS_SEA_LEVEL_PRESSURE_HPA);
-
-        // if we plan to operate in normal mode, set that up and begin sampling
-        // at the specified intervals
-        // if we're going to operate in forced mode, this isn't needed
-        if (_mode == NORMAL_MODE) {
-            // Standby time between samples in normal sampling mode - doesn't
-            // apply in forced mode
-            MS_DBG(F(
-                "Sending BMP3xx stand-by time and starting normal conversion"));
-            bmp_internal.setTimeStandby(_standbyEnum);
-            bmp_internal.startNormalConversion();
+            // if we plan to operate in normal mode, set that up, configure
+            // standby time and filtering, and begin sampling at the specified
+            // intervals if we're going to operate in forced mode, this isn't
+            // needed
+            if (_mode == NORMAL_MODE) {
+                // Coefficient of the filter (in samples)
+                MS_DBG(F("Sending BMP3xx IIR Filter settings"));
+                bmp_internal.setIIRFilter(_filterCoeffEnum);
+                // Standby time between samples in normal sampling mode -
+                // doesn't apply in forced mode
+                MS_DBG(F("Sending BMP3xx stand-by time and starting normal "
+                         "conversion"));
+                bmp_internal.setTimeStandby(_standbyEnum);
+                bmp_internal.startNormalConversion();
+            }
+        } else {
+            MS_DBG(F("Failed to connect to BMP3xx, attempt"), ntries + 1,
+                   F("of 5"));
         }
         ntries++;
     }
@@ -213,9 +218,13 @@ bool BoschBMP3xx::wake(void) {
     // and status bits.  If it returns false, there's no reason to go on.
     if (!Sensor::wake()) return false;
 
-    // if the power has gone off, we need to re-read the coefficients,
-    // we don't need to do anything if always powered.
-    // NOTE:  only forced sampling is supported with switched power
+    // If the power has gone off, we need to re-read the coefficients and
+    // reconfigure the oversampling settings, since the sensor will have lost
+    // its memory of those things.  If the power has not gone off, then the
+    // sensor will still have those things set and we don't need to do anything.
+    // NOTE:  Only forced sampling with the IIR filter disabled is supported
+    // with switched power.  There's no reason to resend the IIR filter and
+    // standby settings because those only apply to continuous power.
     if (_powerPin >= 0) {  // Run begin fxn because it returns true or false
                            // for success in contact
         // Make 5 attempts
@@ -227,11 +236,8 @@ bool BoschBMP3xx::wake(void) {
                   "trim parameters"));
             success = bmp_internal.begin(_i2cAddressHex);
 
-            // Set up oversampling and filter initialization
-            // Using the filter selection recommended for "Weather monitoring
-            // (lowest power)" in table 10 of the sensor datasheet
-
-            // Oversampling setting
+            // Oversampling settings - these settings are sent to the BMP3xx and
+            // need to be repeated at wake if we're not continuously powered
             MS_DBG(F("Sending BMP3xx oversampling settings"));
             bmp_internal.setTempOversampling(_tempOversampleEnum);
             bmp_internal.setPresOversampling(_pressureOversampleEnum);